backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.4.2
python-dotenv==1.0.0
jinja2==3.1.2
mutagen==1.47.0
pillow==10.1.0
starlette==0.27.0
typing-extensions==4.8.0
watchfiles==0.21.0
websockets==11.0.3
tortoise-orm[asyncpg]==0.19.3
